django_notifications/requirements/base.txt
django==3.2.17
uwsgi==2.0.21

python-dotenv==0.21.1
python-dateutil==2.8.2
django-cors-headers==3.13.0
requests==2.28.2
